% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{bootstrap_gene_finder}
\alias{bootstrap_gene_finder}
\title{Function that perform a boostrap process from one cluster over a set of other clusters applying at the end outlier analysis with dbscan.}
\usage{
bootstrap_gene_finder(
  cluster_number,
  seurat_object,
  LFC.lim = 0.2,
  use_limit = FALSE,
  maximo_genes = 20,
  bootstrap.number = 100,
  bootstrap.sample_size = NULL,
  min.p.adj_value = 0,
  ...
)
}
\arguments{
\item{cluster_number}{Cluster number that the function has to process (TODO create a generalization for dataset with cluster with different names)}

\item{seurat_object}{Seural object for single cells that enable us to apply the function FindMarkers}

\item{LFC.lim}{a threshold of log fold change when selecting genes as input to perform Wilcoxon's test.}

\item{use_limit}{TRUE/FALSE. If the process selects just a fixed number of genes.}

\item{maximo_genes}{Maximum number of genes that have to be selected.}

\item{bootstrap.number}{How many boostraping loops the algorithm is going to execute}

\item{min.p.adj_value}{Value that should be greater or equal to zero. Normally it is 0.05}
}
\value{
List with the marker genes
}
\description{
Function that perform a boostrap process from one cluster over a set of other clusters applying at the end outlier analysis with dbscan.
}
