% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{calculate_foldchange}
\alias{calculate_foldchange}
\title{Function that calculates the foldchange between a given cluster over the rest}
\usage{
calculate_foldchange(sc.eset, ct.varname, ct.group.temp, LFC.lim)
}
\arguments{
\item{sc.eset}{ExpressionSet object for single cell samples}

\item{ct.varname}{variable name for 'cell types'}

\item{LFC.lim}{a threshold of log fold change when selecting genes as input to perform Wilcoxon's test.}

\item{bulk.eset}{ExpressionSet object for bulk samples}
}
\value{
Estimated proportion, basis matrix, predicted gene expression levels for bulk samples
}
\description{
Function that calculates the foldchange between a given cluster over the rest (It doesn't use an external library to apply the wilcox algorithm). At the end it uses bonferroni correction
}
