% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{CreateSCDCpropObj}
\alias{CreateSCDCpropObj}
\title{Create user-defined SCDC_prop object, as the SCDC_ENSEMBLE input}
\usage{
CreateSCDCpropObj(p, b)
}
\arguments{
\item{p}{the estimated proportion matrix. sample by cell type.}

\item{b}{the estimated signature/feature/basis matrix. gene by cell type.}
}
\description{
Create user-defined SCDC_prop object, as the SCDC_ENSEMBLE input
}
