% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{custom_apply}
\alias{custom_apply}
\title{custom_apply Function}
\arguments{
\item{X}{matrix to apply the function (sparse matrix)}

\item{MARGIN}{1 for rows, 2 for columns}

\item{FUN}{function for apply to the X matrix}
}
\value{
Matrix resulted with the function applied to original X matrix
}
\description{
Apply function for sparse matrix with the same parameter form that the native apply.
}
\author{
Crhisto
}
