% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/densityCI.R
\name{densityCI}
\alias{densityCI}
\title{densityCI}
\usage{
densityCI(res, condition)
}
\arguments{
\item{res}{results from \code{scDC_noClustering}, \code{scDC_clustering} function}

\item{condition}{a vector indicate the condition associate the results}
}
\value{
Returns a data frame.
}
\description{
Visualising the CI using density plots of proportions
}
\examples{
## Loading example data

data("sim")

cellTypes = sim$sim_cellTypes
subject = sim$sim_subject
\dontrun{
res_scDC = scDC_noClustering(cellTypes, subject,
calCI = TRUE, calCI_method = c("BCa", "percentile"))
densityCI(res_scDC, c("cond1","cond1","cond2","cond2"))

}

}
\author{
Yingxin Lin
}
