% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSEMBLE.R
\name{ENSEMBLE_prop}
\alias{ENSEMBLE_prop}
\title{calculate weighted proportions after ENSEMBLE step}
\usage{
ENSEMBLE_prop(ENSobject, weight_method = "mAD_Y_LAD", other_wts = NULL)
}
\arguments{
\item{ENSobject}{The object from the ENSEMBLE result, either from SCDC_ENSEMBLE_subcl() or SCDC_ENSEMBLE() or you can create one by yourself: list(w_table=, prop.list=), where w_table contains weights for each reference dataset, and prop.list is a list of deconvolution results (with prop.est)}

\item{weight_method}{Specify which set of weights you want to use to ENSEMBLE. Suggested are "mAD_Y_LAD", "min.mAD_Y", "Spearman", "NNLS". If not using the suggested methods, you can input your own designed weights.}

\item{other_wts}{Specify your own designed weights. If you use the suggested methods, leave this as NULL value.}
}
\description{
ENSEMBLE function for manually input deconvolution results
}
