% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitGLM.R
\name{fitGLM}
\alias{fitGLM}
\title{fitGLM}
\usage{
fitGLM(res, condition, subject_effect = TRUE, pairwise = TRUE,
  fixed_only = FALSE, verbose = TRUE)
}
\arguments{
\item{res}{results from \code{scDC_noClustering}, \code{scDC_clustering} function}

\item{condition}{a vector indicate the condition associate the results}

\item{subject_effect}{A logical input for whether fit the subject effect}

\item{pairwise}{A logical input for whether the subject in different condition are paired}

\item{fixed_only}{A logical input for whether only fiftting the fixed effect GLM model.}

\item{verbose}{A logical input for whether print the progress.}
}
\value{
return GLM results
}
\description{
Visualising the CI using barplot
}
\examples{
## Loading example data

\dontrun{
library(scDC)
data("sim")

cellTypes = sim$sim_cellTypes
subject = sim$sim_subject
res_scDC = scDC_noClustering(cellTypes, subject,
calCI = TRUE, calCI_method = c("BCa", "percentile"))
barplotCI(res_scDC, c("cond1","cond1","cond2","cond2"))

}

}
\author{
Yingxin Lin
}
