% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{generateBulk_allcells}
\alias{generateBulk_allcells}
\title{Construct Pseudo bulk samples}
\usage{
generateBulk_allcells(eset, ct.varname, sample, disease = NULL,
  ct.sub = NULL)
}
\arguments{
\item{eset}{ExpressionSet object for single cells}

\item{ct.varname}{variable name for 'cell types'}

\item{sample}{variable name for subject/samples}

\item{disease}{indicate the health condition of subjects}

\item{ct.sub}{a subset of cell types that are selected to construct pseudo bulk samples. If NULL, then all cell types are used.}
}
\value{
pseudo bulk samples ExpressionSet, and actual cell-type proportions
}
\description{
Construct Pseudo bulk samples by actual number of cells per subject
}
