% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{getESET}
\alias{getESET}
\title{Get ExpressionSet}
\arguments{
\item{exprs}{raw count matrix}

\item{fdata}{feature data, for genes, usually it's gene name}

\item{pdata}{pheno data, for samples, usually it's the characteristics for each single cell/bulk sample, including name, gender, age, cluster, disease,...}
}
\description{
Use Pdata, Fdata, and count matrix to derive ExpressionSet Object
}
