% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{getSearchGrid}
\alias{getSearchGrid}
\title{Grid search matrix}
\usage{
getSearchGrid(lengthby, nparam)
}
\arguments{
\item{lengthby}{search length}

\item{nparam}{number of paramters/reference datasets}
}
\description{
Generate Grid-search matrix
}
