% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{iteration_over_clusters_default_parallelized}
\alias{iteration_over_clusters_default_parallelized}
\title{parallelization of the for the original procedure with some additional lines for:}
\usage{
iteration_over_clusters_default_parallelized(
  ct.group,
  LFC.lim,
  sc.eset,
  pseudocount.use,
  core_number = NULL,
  ...
)
}
\arguments{
\item{ct.group}{List of clusters that will be analyzed}

\item{LFC.lim}{Foldchange limit for the comparison between each cluster amoung the others}

\item{sc.eset}{ExpressionSet object for single cells}

\item{core_number}{Number of cores that will be used for the process.}
}
\value{
List with the marker genes for all selected clusters
}
\description{
parallelization of the for the original procedure with some additional lines for: 1. Show the selected genes. 2. Add parallelization. 3. Add logging function in a temporal file.
}
