% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{iteration_over_clusters_parallelized}
\alias{iteration_over_clusters_parallelized}
\alias{qc_iteration_parallelized}
\title{Paralellization of the for procedure for QC process}
\usage{
qc_iteration_parallelized(
  m.basis,
  xsc,
  sigma,
  N.sc,
  nu,
  epsilon,
  iter.max,
  core_number = NULL,
  ...
)

iteration_over_clusters_parallelized(
  ct.group,
  LFC.lim,
  sc.eset,
  top_genes,
  pseudocount.use,
  core_number = NULL,
  fix_number_genes = NULL,
  ...
)
}
\arguments{
\item{nu}{a small constant to facilitate the calculation of variance}

\item{epsilon}{a small constant number used for convergence criteria}

\item{iter.max}{the maximum number of iteration in WNNLS}

\item{core_number}{Number of cores that will be used for the process.}

\item{ct.group}{List of clusters that will be analyzed}

\item{LFC.lim}{a threshold of log fold change when selecting genes as input to perform Wilcoxon's test.}

\item{sc.eset}{ExpressionSet object for single cells}

\item{top_genes}{Number of top genes that we want to consider to check the maximum LFC for each cluster. This is just for informative purposes.}

\item{fix_number_genes}{Maximum number of marker genes that has to be returned by cluster}
}
\value{


List with the marker genes for all selected clusters
}
\description{
It contains some experiments but is not the final version. It is just saved for documentation purposes.
}
