% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{iteration_over_clusters_parallelized_wilcox}
\alias{iteration_over_clusters_parallelized_wilcox}
\alias{iteration_over_clusters_parallelized_wilcox_outlier}
\title{Marker gene selection with parallelization using wilcox method from the Seurat library}
\usage{
iteration_over_clusters_parallelized_wilcox_outlier(
  sc.eset,
  ct.group,
  core_number = NULL,
  LFC.lim = 0.2,
  minimun_number_markers = 20,
  minimum_genes_pipeline = TRUE,
  add_zero_p_val_adj = TRUE,
  minimum_p_val_adj = 0,
  maximo_genes = 20,
  ...
)
}
\arguments{
\item{sc.eset}{ExpressionSet object for single cells}

\item{ct.group}{List of clusters that will be analyzed}

\item{core_number}{Number of cores that will be used for the process.}

\item{LFC.lim}{a threshold of log fold change when selecting genes as input to perform Wilcoxon's test.}

\item{minimun_number_markers}{If the variable minimum_genes_pipeline is active and also if in the first step with the wilcox text there are at least this genes, the process is not going to use outlier detection with dbscan}

\item{minimum_genes_pipeline}{TRUE/FALSE. Parameter that enable the minimum gene process using outlier detection. If is FALSE the process is not going to run and the genes would be the ones found for wilcox test}

\item{add_zero_p_val_adj}{TRUE/FALSE. If after the dbscan we want to add the genes that were found with the normal wilcox detection using the threshold (normally 0 or 0.05)}

\item{minimum_p_val_adj}{p-value-adjusted for the wilcox process. Could be zero or 0.05}

\item{maximo_genes}{If we want to limitate the maximum number of marker genes at the end of the process.}
}
\value{
List with the marker genes for all selected clusters
}
\description{
Marker gene selection with parallelization using wilcox method from the Seurat library
}
