% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{iteration_over_clusters_parallelized_wilcox_boostrap}
\alias{iteration_over_clusters_parallelized_wilcox_boostrap}
\title{Marker gene selection with parallelization using wilcox method from the Seurat library}
\usage{
iteration_over_clusters_parallelized_wilcox_boostrap(
  sc.eset,
  ct.group,
  core_number = NULL,
  LFC.lim = 0.2,
  iteration.minimun_number_markers = 28,
  iteration.use_maximum = TRUE,
  iteration.maximo_genes = 35,
  iteration.use_final_foldchange = FALSE,
  bootstrap.sample_size = NULL,
  bootstrap.number = NULL,
  ...
)
}
\arguments{
\item{sc.eset}{ExpressionSet object for single cells}

\item{ct.group}{List of clusters that will be analyzed}

\item{core_number}{Number of cores that will be used for the process.}

\item{LFC.lim}{Fa threshold of log fold change when selecting genes as input to perform Wilcoxon's test.}

\item{iteration.use_final_foldchange}{TRUE/FALSE. If at the end the cluster has zero genes if this parameter is true, the boostraping is going to be calculated over the foldchange with <0.05, not with zero.}

\item{minimun_number_markers}{If in the first step with the wilcox text there are at least this genes, the process is not going to use outlier detection with dbscan}

\item{use_maximum}{TRUE/FALSE. If the process selects just a fixed number of genes.}

\item{maximo_genes}{If we want to limitate the maximum number of marker genes at the end of the process.}
}
\value{
List with the marker genes for all selected clusters
}
\description{
Marker gene selection with parallelization using wilcox method from the Seurat library
}
