% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{my_qc}
\alias{my_qc}
\title{Clustering QC}
\usage{
my_qc(sc.eset, ct.varname, sample, scsetname = "Single Cell", ct.sub,
  iter.max = 1000, nu = 1e-04, epsilon = 0.01, arow = NULL,
  qcthreshold = 0.7, ...)
}
\arguments{
\item{sc.eset}{ExpressionSet object for single cells}

\item{ct.varname}{variable name for 'cell type'}

\item{sample}{variable name for subject/sample}

\item{scsetname}{the name for the single cell dataset}

\item{ct.sub}{a subset of cell types that are selected to construct basis matrix}

\item{iter.max}{the maximum number of iteration in WNNLS}

\item{nu}{a small constant to facilitate the calculation of variance}

\item{epsilon}{a small constant number used for convergence criteria}

\item{arow}{annotation of rows for pheatmap}

\item{qcthreshold}{the probability threshold used to filter out questionable cells}
}
\value{
a list including: 1) a probability matrix for each single cell input; 2) a clustering QCed ExpressionSet object; 3) a heatmap of QC result.
}
\description{
Single cells Clustering QC
}
