% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{save_log_file}
\alias{save_log_file}
\title{Function to save message in a log file for each loop:}
\usage{
save_log_file(file_path, file_name, message1, message2)
}
\arguments{
\item{file_path}{Path in the server to save the file}

\item{file_name}{Name of the file to be saved}

\item{message1}{Message # 1 to be saved}
}
\description{
Function to save message in a log file for each loop
}
