% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{SCDC_basis}
\alias{SCDC_basis}
\title{Basis Matrix}
\usage{
SCDC_basis(x, ct.sub = NULL, ct.varname, sample)
}
\arguments{
\item{x}{ExpressionSet object for single cells}

\item{ct.sub}{a subset of cell types that are selected to construct basis matrix}

\item{ct.varname}{variable name for 'cell types'}

\item{sample}{variable name for subject/samples}
}
\value{
a list of basis matrix, sum of cell-type-specific library size, sample variance matrix, basis matrix by mvw, mvw matrix.
}
\description{
Basis matrix construction
}
