% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSEMBL.R
\name{SCDC_ENSEMBLE}
\alias{SCDC_ENSEMBLE}
\title{ENSEMBL function}
\usage{
SCDC_ENSEMBLE(bulk.eset, sc.eset.list, ct.varname, sample, ct.sub,
  search.method = c("Grid search", "LAD", "NNLS"),
  search.length = 0.05, iter.max = 2000, nu = 1e-04,
  epsilon = 0.001, truep = NULL, ...)
}
\arguments{
\item{bulk.eset}{ExpressionSet object for bulk samples}

\item{sc.eset.list}{list of ExpressionSet objects for single cell reference datasets}

\item{ct.varname}{variable name for 'cell types'}

\item{sample}{variable name for subject/samples}

\item{ct.sub}{a subset of cell types that are selected to construct basis matrix}

\item{search.method}{method to derive the ENSEMBL weights, including "Grid search", "LAD", "NNLS".}

\item{search.length}{if using "Grid search", the step length used.}

\item{iter.max}{the maximum number of iteration in WNNLS}

\item{nu}{a small constant to facilitate the calculation of variance}

\item{epsilon}{a small constant number used for convergence criteria}

\item{truep}{true cell-type proportions for bulk samples if known}
}
\description{
ENSEMBL function for deconvolution results output from SCDC_prop
}
