% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ENSEMBLE.R
\name{SCDC_ENSEMBLE_subcl}
\alias{SCDC_ENSEMBLE_subcl}
\title{ENSEMBL function for manually input deconvolution results}
\usage{
SCDC_ENSEMBLE_subcl(bulk.eset, prop.list, truep = NULL, ct.sub,
  search.length = 0.05, search.method = c("Grid search", "LAD",
  "NNLS"), iter.max = 2000, nu = 1e-04, epsilon = 0.001, ...)
}
\arguments{
\item{bulk.eset}{ExpressionSet object for bulk samples}

\item{prop.list}{list of deconvolution results}

\item{truep}{true cell-type proportions for bulk samples if known}

\item{ct.sub}{a subset of cell types that are selected to construct basis matrix}

\item{search.length}{if using "Grid search", the step length used.}

\item{search.method}{method to derive the ENSEMBL weights, including "Grid search", "LAD", "NNLS".}

\item{iter.max}{the maximum number of iteration in WNNLS}

\item{nu}{a small constant to facilitate the calculation of variance}

\item{epsilon}{a small constant number used for convergence criteria}

\item{ct.varname}{variable name for 'cell types'}

\item{sample}{variable name for subject/samples}
}
\description{
ENSEMBL function for manually input deconvolution results
}
