% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{SCDC_peval}
\alias{SCDC_peval}
\title{Evaluate deconvolved proportions}
\usage{
SCDC_peval(ptrue, pest, pest.names, select.ct = NULL)
}
\arguments{
\item{ptrue}{a matrix of true/actual cell-type proportions for bulk samples}

\item{pest}{a list of estimated cell-type proportion matrix}

\item{pest.names}{method name for the estimated proportions in the pest list}

\item{select.ct}{selected cell types for deconvolution}
}
\description{
Evaluation function, for deconvolved proportions and the actual/true proportions
}
