% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Deconvolution.R
\name{SCDC_qc_ONE}
\alias{SCDC_qc_ONE}
\title{Clustering QC for single cells from one subject}
\usage{
SCDC_qc_ONE(
  sc.eset,
  ct.varname,
  sample,
  scsetname = "Single Cell",
  ct.sub,
  iter.max = 1000,
  nu = 1e-04,
  epsilon = 0.01,
  arow = NULL,
  weight.basis = F,
  qcthreshold = 0.7,
  generate.figure = T,
  cbPalette = c("#999999", "#E69F00", "#56B4E9", "#009E73", "#F0E442", "#0072B2",
    "#D55E00", "#CC79A7"),
  ...
)
}
\arguments{
\item{sc.eset}{ExpressionSet object for single cells}

\item{ct.varname}{variable name for 'cell type'}

\item{sample}{variable name for subject/sample}

\item{scsetname}{the name for the single cell dataset}

\item{ct.sub}{a subset of cell types that are selected to construct basis matrix}

\item{iter.max}{the maximum number of iteration in WNNLS}

\item{nu}{a small constant to facilitate the calculation of variance}

\item{epsilon}{a small constant number used for convergence criteria}

\item{arow}{annotation of rows for pheatmap}

\item{qcthreshold}{the probability threshold used to filter out questionable cells}

\item{generate.figure}{logical. If generate the heatmap by pheatmap or not. default is TRUE.}
}
\value{
a list including: 1) a probability matrix for each single cell input; 2) a clustering QCed ExpressionSet object; 3) a heatmap of QC result.
}
\description{
Clustering QC for single cells from one subject
}
