% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{SCDC_yeval}
\alias{SCDC_yeval}
\title{Evaluate predicted gene expression levels}
\usage{
SCDC_yeval(y, yest, yest.names = NULL)
}
\arguments{
\item{y}{observed raw counts of a bulk sample}

\item{yest}{a list of predicted gene expression levels}

\item{yest.names}{method name for the predicted gene expression levels in yest list}
}
\description{
Evaluation function, for observed and predicted gene expression levels
}
