% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sim}
\alias{sim}
\title{simulated expression matrix of single cell RNA-seq data, with cell type, subject, and condition info}
\format{A `list` object}
\usage{
sim
}
\description{
A dataset simulated from PowsimR
sim
}
\keyword{datasets}
