% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Visualization.R
\name{wheat_map}
\alias{wheat_map}
\title{Heatmap for ENSEMBL weight selection, with four measures}
\usage{
wheat_map(ensemble_res, ref1, ref2)
}
\arguments{
\item{ensemble_res}{the EMSEMBL result object, derived from SCDC_ENSEMBL() or SCDC_ENSEMBL_subcl()}

\item{ref1}{name for the first reference dataset}

\item{ref2}{name for the second reference dataset}
}
\value{
A figure of evaluated performance, varying the ENSEMBLE weights for three reference datasets
}
\description{
Heatmap for ENSEMBL weight selection, with four measures
}
