% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Basic_Functions.R
\name{wt_y}
\alias{wt_y}
\title{Calculate cell type proportions by suggested weights}
\usage{
wt_y(wt, y.list = y.list)
}
\arguments{
\item{wt}{a vector of weights for each reference dataset. Should be of the same length as the y.list.}

\item{y.list}{a matrix, with each column containing the vectorized estimated gene expressions from each reference dataset.}
}
\description{
Calculate proportions by linear combination of a list of proportions
}
